pyproject.toml
[tool.poetry]
name = "graphdb-access-security-hub"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
faststream = {extras = ["rabbit"], version = "^0.4.3"}
pyneo4j-ogm = ">=0.4.1"
pyyaml = "^6.0.1"
aiofiles = "^23.2.1"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.1"
mypy = "^1.8.0"
ruff = "^0.1.14"
testcontainers = "^3.7.1"
testcontainers-neo4j = "^0.0.1rc1"
pytest-cov = "^4.1.0"
types-aiofiles = "^23.2.0.20240106"
types-pyyaml = "^6.0.12.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
